+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Change log
 
+## 0.5.0 (2023-10-10)
+
+### Features
+
+- Introduced artists filter widget on the bandcamp label page
+- Artists filter widget is synchronized with the popup releases data widget
+
 ## 0.4.1 (2023-09-29)
 
 ### Fixed